refactor: Clarify file content
